Fixed build failure when compiling with clang++
src/LuaPlus/LuaPlusCD.h:1045:61: error: cast from pointer to smaller type 'unsigned int' loses information
            LPCD::Push(L, *(VarType*)((unsigned char*)obj + (unsigned int)offset));
                                                            ^~~~~~~~~~~~~~~~~~~~
src/LuaPlus/LuaPlusCD.h:1059:47: error: cast from pointer to smaller type 'unsigned int' loses information
            *(VarType*)((unsigned char*)obj + (unsigned int)offset) = LPCD::Get(LPCD::TypeWrapper<VarType>(), L, 2);
                                              ^~~~~~~~~~~~~~~~~~~~

Even Microsoft supports 'stdint.h' in the latest versions of their
compiler nowadays, so it's time to use it safely.

From: Siarhei Siamashka <siarhei.siamashka@gmail.com>

git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@1167 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/LuaPlus/LuaPlusCD.h
@@ -26,6 +26,7 @@ extern "C" {
 
 #include <typeinfo>
 #include <string.h>
+#include <stdint.h>
 
 // LuaPlus Call Dispatcher
 namespace LPCD
@@ -1042,7 +1043,7 @@ namespace LPCD
 
             Object* obj = (Object*)LPCD::GetObjectUserData(L);
 
-            LPCD::Push(L, *(VarType*)((unsigned char*)obj + (unsigned int)offset));
+            LPCD::Push(L, *(VarType*)((unsigned char*)obj + (uintptr_t)offset));
 
             return 1;
         }
@@ -1056,7 +1057,7 @@ namespace LPCD
             if (!Match(TypeWrapper<VarType>(), L, 2))
                 luaL_argerror(L, 2, "bad argument");
 
-            *(VarType*)((unsigned char*)obj + (unsigned int)offset) = LPCD::Get(LPCD::TypeWrapper<VarType>(), L, 2);
+            *(VarType*)((unsigned char*)obj + (uintptr_t)offset) = LPCD::Get(LPCD::TypeWrapper<VarType>(), L, 2);
 
             return 1;
         }
